Add backoff-retry delays between block upload attempts
There's more potential for error when uploading a snapshot.  For example, we've
seen 4 "connection closed" errors in a row for a block upload.  The most
pernicious is the "snapshot does not exist" error, when there are timing issues
that cause new snapshots to not be immediately available for block upload,
which can take a couple minutes to pass.