Fix the wrong type i32 used for offset calculation at download to u64
The offset calculation for writing a block to download target file is
in i32 after commit 118e5861, which replaced the AWS SDK from rusoto
to AWS's one and also replaced the type of block indext and block size.
This causes an integer overflow when there is a block whoes index is
greater than 4096 or the offset is greater than 2 GiB as the block size
is 512 KiB.

This commit will change the type casting of the offset to type castings
of block indext and block size to prevent integer overflows at offset
calculations for downloads.