Add debug logging to help with troubleshooting
This adds a `--verbose` flag to the CLI to enable debug level log
output. This can be useful when something is failing or taking a long
time. With normal execution there is no indication of what is happening.
This will at least give some sort of troubleshooting step a user can
take to help triage what is happening.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>