Allow users to provide tags on upload
This change adds the ability for users to pass in any number of
tags during upload. The tags are passed as given to EBS start-snapshot.
Tags are expected to be inputted as "Key=k,Value=v" where k and v
are users desired tags. The key and value delimiters were chosen
to adhere as much as possible to other AWS cli tag inputs. Any
restrictions on tags are expected to be handled by the underlining
EBS apis.